Add --no-problem-statement and --only-last-submission flags (#2)
* Add --no-problem-statement and --only-last-submission flags

* Update script help message in README.md

---------

+++ leetcode_export/__main__.py
@@ -31,6 +31,12 @@ def parse_args():
         help="problem folder name format",
     )
     parser.add_argument(
+        "--no-problem-statement",
+        dest="no_problem_statement",
+        action="store_true",
+        help="do not save problem statement",
+    )
+    parser.add_argument(
         "--problem-statement-filename",
         type=str,
         default="${question_id}-${title_slug}.html",
@@ -54,6 +60,12 @@ def parse_args():
         dest="only_accepted",
         action="store_true",
         help="save accepted submissions only",
+    )
+    parser.add_argument(
+        "--only-last-submission",
+        dest="only_last_submission",
+        action="store_true",
+        help="only save the last submission for each programming language",
     )
     parser.add_argument(
         "--language",
@@ -145,13 +157,40 @@ def main():
     os.chdir(args.folder)
 
     title_slug_to_problem_folder_name: dict[str, str] = dict()
+    title_slug_to_exported_languages: dict[str, set[str]] = dict()
+
+    last_submission_timestamp: Optional[int] = None
 
     for submission in leetcode.get_submissions():
+        if (
+            last_submission_timestamp is not None
+            and submission.timestamp > last_submission_timestamp
+        ):
+            logging.warning(
+                "Submissions are not in reverse chronological order, --only-last-submission flag might not work as expected if used. Please report this issue on GitHub attaching the debug.log file: https://github.com/NeverMendel/leetcode-export/issues"
+            )
+        last_submission_timestamp = submission.timestamp
+
         if args.only_accepted and submission.status_display != "Accepted":
             continue
 
         if args.language and submission.lang not in args.language:
             continue
+
+        if submission.title_slug not in title_slug_to_exported_languages:
+            title_slug_to_exported_languages[submission.title_slug] = set()
+
+        if (
+            args.only_last_submission
+            and submission.title_slug in title_slug_to_exported_languages
+            and submission.lang
+            in title_slug_to_exported_languages[submission.title_slug]
+        ):
+            logging.info(
+                f"The latest submission for {submission.title_slug} in {submission.lang} has already been exported, skipping this submission"
+            )
+            continue
+        title_slug_to_exported_languages[submission.title_slug].add(submission.lang)
 
         if submission.title_slug not in title_slug_to_problem_folder_name:
             problem_statement = leetcode.get_problem_statement(submission.title_slug)
@@ -168,7 +207,9 @@ def main():
             problem_statement_filename = problem_statement_filename_template.substitute(
                 **problem_statement.__dict__
             )
-            if not os.path.exists(problem_statement_filename):
+            if not args.no_problem_statement and not os.path.exists(
+                problem_statement_filename
+            ):
                 with open(problem_statement_filename, "w+") as problem_statement_file:
                     problem_statement_file.write(
                         problem_statement_template.substitute(
